feat: apply v2 data structures to existing codes

+++ data/src/commonMain/kotlin/xyz/hyli/timeflow/data/v2/Date.kt
@@ -51,7 +51,7 @@ data class Date(
             }"
         }
 
-    fun toLocalDate(): LocalDate = LocalDate(year.toInt(), month.toInt(), day.toInt())
+    fun toLocalDate(): LocalDate = LocalDate(year, month, day)
 
     @OptIn(ExperimentalTime::class)
     fun weeksTill(date: LocalDate = Clock.System.todayIn(TimeZone.currentSystemDefault())): Int {
